refactor(controller): extract llama.cpp arg appenders to runtime_llamacpp_args.go
Part 1/7 of splitting the 1,567-line inferenceservice_controller.go
into per-concern files. Tracks #311.

Move fourteen llama.cpp argument-builder helpers from the god file
into a new runtime_llamacpp_args.go sibling (alongside the existing
runtime_llamacpp.go). Pure relocation — no signature changes, no
behavior changes, no callers updated (callers are already in the same
package and still resolve the unqualified names).

Moved functions:
  appendContextSizeArgs, appendParallelSlotsArgs, appendFlashAttentionArgs,
  appendJinjaArgs, appendCacheTypeArgs, appendMoeCPUOffloadArgs,
  appendMoeCPULayersArgs, appendNoKvOffloadArgs, appendTensorOverrideArgs,
  appendMetadataOverrideArgs, appendBatchSizeArgs, appendUBatchSizeArgs,
  appendNoWarmupArgs, appendReasoningBudgetArgs

Line deltas:
  inferenceservice_controller.go: 1,567 → 1,463 (−104)
  runtime_llamacpp_args.go:                   0 → 130 (new file)

Verified: go vet clean, golangci-lint 0 issues, make test green,
internal/controller coverage unchanged at 83.1%.
